// routes/logic/classes/interfaces/UserSoasInterface.ts
export interface UserSoasInterface {
    userSoasName: string;
    userSoasLogin: string;
    userSoasPassword: string;
    userSoasRole: string;
    userSoasLanguage: string;
}

export interface UserSoasDataInterface {
    USER_SOAS_ID: string;
    USER_LANGUAGE: string;
    USER_SOAS_LOGIN: string;
    USER_SOAS_PASSWD: string;
    USER_SOAS_ROLE: string;
}